Revert "LPS-187086 SF, move to package that matches the interface it implements"
This reverts commit 3f96059b08445dcc4ff85b8eb606358623bf8e3e.
