+++ modules/apps/oauth2-provider/oauth2-provider-scope-impl/src/main/java/com/liferay/oauth2/provider/scope/internal/ThreadLocalScopeContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.oauth2.provider.scope.internal.liferay;
+package com.liferay.oauth2.provider.scope.internal;
 
 import com.liferay.oauth2.provider.scope.liferay.ScopeContext;
 import com.liferay.petra.string.StringPool;
